Split project and cluster handler
